Change correction of UP axis to rotation
The old method for correcting UP axis switches left and right so I
changed it to a rotation. Metasequoia has CW winding for front whereas
Blender uses CCW which means faces must always be inverted. The default
for Correction  of inverted faces proprty is now True. The old method
for correcting UP axis inverted faces automatically. The UVs were not
adjusted if the faces were inverted so I had to add code for that case.